requirements.txt
absl-py==0.7.1
alabaster==0.7.12
albumentations==0.3.0
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.2
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
astropy==3.1.2
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.8.3
bkcharts==0.2
bleach==3.1.0
blis==0.4.1
bokeh==1.0.4
boto==2.49.0
Bottleneck==1.2.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
clyent==1.2.2
cmake==3.20.3
colorama==0.4.1
comtypes==1.1.7
conda-package-handling==1.3.10
conda-verify==3.1.1
contextlib2==0.5.5
cryptography==2.6.1
ctk-cli==1.5
cycler==0.10.0
cymem==2.0.2
Cython==0.29.6
cytoolz==0.9.0.1
dask==1.1.4
decorator==4.4.0
defusedxml==0.5.0
diskcache==5.2.1
distributed==1.26.0
distro==1.5.0
docker==5.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
efficientnet==0.0.4
efficientnet-pytorch==0.4.0
entrypoints==0.3
et-xmlfile==1.0.1
fastai==1.0.59
fastcache==1.0.2
fastprogress==0.1.21
filelock==3.0.10
Flask==1.0.2
future==0.17.1
gast==0.2.2
gevent==1.4.0
girder-client==3.1.5
glob2==0.6
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.22.0
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
imgaug==0.2.6
importlib-metadata==0.23
imutils==0.5.2
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.16
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-resnet==0.1.0
keyring==18.0.0
kiwisolver==1.0.1
large-image==1.6.1
large-image-source-pil==1.6.1
large-image-source-tiff==1.6.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
libtiff==0.4.2
livelossplot==0.4.1
llvmlite==0.34.0
lmdb==0.97
locket==0.2.0
lxml==4.3.2
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mock==3.0.5
more-itertools==6.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
murmurhash==1.0.2
navigator-updater==0.2.1
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nibabel==3.2.0.dev0
nimfa==1.4.0
ninja==1.10.0.post2
nltk==3.4
nose==1.3.7
notebook==5.7.8
numba==0.51.2
numexpr==2.6.9
numpy==1.19.2
numpydoc==0.8.0
nvidia-ml-py3==7.352.0
olefile==0.46
opencv-python==4.1.0.25
opencv-python-headless==4.1.0.25
openpyxl==2.6.1
openslide-python==1.1.1
opt-einsum==3.3.0
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
partd==0.3.10
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==8.2.0
pkginfo==1.5.0.1
plac==1.1.3
pluggy==0.9.0
ply==3.11
preshed==3.0.2
progressbar2==3.42.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.6.1
py==1.8.0
py-wsi==2.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycuda==2019.1.2
pycurl==7.43.0.2
pydicom==1.3.0
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.3.1
pykwalify==1.7.0
pylint==2.3.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyradiomics==3.0.1
pyreadline==2.1
pyrsistent==0.14.11
PySocks==1.6.8
pytest==4.3.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
python-dotenv==0.17.1
python-telegram-bot==12.0.0b1
python-utils==2.3.0
pytools==2019.1.1
pytorch-gradcam==0.2.1
pytorchcv==0.0.50
pytz==2018.9
pyvips==2.1.14
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.7.0
requests==2.21.0
requests-toolbelt==0.9.1
rope==0.12.0
ruamel-yaml==0.15.46
scikit-build==0.11.1
scikit-image==0.18.1
scikit-learn==0.20.3
scikit-multilearn==0.2.0
scipy==1.5.2
seaborn==0.11.1
segmentation-models==1.0.1
Send2Trash==1.5.0
session-info==1.0.0
setuptools-scm==6.0.1
Shapely==1.6.4.post2
simplegeneric==0.8.1
SimpleITK==1.2.0
singledispatch==3.4.0.3
six==1.12.0
slideio==0.4.100
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
spacy==2.2.2
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.3.1
srsly==0.2.0
staintools==2.1.2
statannot==0.2.2
statsmodels==0.9.0
stdlib-list==0.8.0
sympy==1.3
tables==3.5.1
tblib==1.3.2
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
texttable==1.6.3
thinc==7.3.1
tifffile==2021.6.14
toolz==0.9.0
torch==1.2.0
torchvision==0.4.0
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
umap-learn==0.4.6
unicodecsv==0.14.1
urllib3==1.24.1
virtualenv==16.6.1
wasabi==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.1
xgboost==1.4.2
xlrd==1.2.0
XlsxWriter==1.1.5
xlwings==0.15.4
xlwt==1.3.0
yellowbrick==1.3.post1
zict==0.1.4
zipp==0.6.0
